api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
google-genai>=1.0.0
psycopg2-binary>=2.9
pgvector>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
